# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.AddCommentReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      mention_username_list:
        description: '@ user id list'
        items:
          type: string
        type: array
      object_id:
        description: object id
        type: string
      original_text:
        description: original comment content
        maxLength: 600
        minLength: 2
        type: string
      reply_comment_id:
        description: reply comment id
        type: string
    required:
    - object_id
    - original_text
    type: object
  schema.AddReportReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        description: captcha_id
        type: string
      content:
        description: report content
        maxLength: 500
        type: string
      object_id:
        description: object id
        maxLength: 20
        type: string
      report_type:
        description: report type
        type: integer
    required:
    - object_id
    - report_type
    type: object
...
  schema.GetCommentPersonalWithPageResp:
    properties:
      answer_id:
        description: answer id
        type: string
      comment_id:
        description: comment id
        type: string
      content:
        description: content
        type: string
      created_at:
        description: create time
        type: integer
      object_id:
        description: object id
        type: string
      object_type:
        description: object type
        enum:
        - question
        - answer
        - tag
        - comment
        type: string
      question_id:
        description: question id
        type: string
      title:
        description: title
        type: string
      url_title:
        description: url title
        type: string
    type: object
  schema.GetCommentResp:
    properties:
      comment_id:
        description: comment id
        type: string
      created_at:
        description: create time
        type: integer
      is_vote:
        description: current user if already vote this comment
        type: boolean
      member_actions:
        description: MemberActions
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      object_id:
        description: object id
        type: string
      original_text:
        description: original comment content
        type: string
      parsed_text:
        description: parsed comment content
        type: string
      reply_comment_id:
        description: reply comment id
        type: string
      reply_user_display_name:
        description: reply user display name
        type: string
      reply_user_id:
        description: reply user id
        type: string
      reply_user_status:
        description: reply user status
        type: string
      reply_username:
        description: reply user username
        type: string
      user_avatar:
        description: user avatar
        type: string
      user_display_name:
        description: user display name
        type: string
      user_id:
        description: user id
        type: string
      user_status:
        description: user status
        type: string
      username:
        description: username
        type: string
      vote_count:
        description: user vote amount
        type: integer
    type: object
...
  schema.GetReportListPageResp:
    properties:
      answer_accepted:
        type: boolean
      answer_count:
        type: integer
      answer_id:
        type: string
      author_user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      comment_id:
        type: string
      created_at:
        type: integer
      flag_id:
        type: string
      object_id:
        type: string
      object_show_status:
        type: integer
      object_status:
        type: integer
      object_type:
        enum:
        - question
        - answer
        - comment
        type: string
      original_text:
        type: string
      parsed_text:
        type: string
      question_id:
        type: string
      reason:
        $ref: '#/definitions/schema.ReasonItem'
      reason_content:
        type: string
      submit_at:
        type: integer
      submitter_user:
        $ref: '#/definitions/schema.UserBasicInfo'
      tags:
        items:
          $ref: '#/definitions/schema.TagResp'
        type: array
      title:
        type: string
      url_title:
        type: string
    type: object
...
  schema.PermissionMemberAction:
    properties:
      action:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  schema.ReasonItem:
    properties:
      content_type:
        type: string
      description:
        type: string
      name:
        type: string
      placeholder:
        type: string
      reason_key:
        type: string
      reason_type:
        type: integer
    type: object
...
  schema.RemoveCommentReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      comment_id:
        description: comment id
        type: string
    required:
    - comment_id
    type: object
...
  schema.ReviewReportReq:
    properties:
      close_msg:
        type: string
      close_type:
        type: integer
      content:
        maxLength: 65535
        minLength: 6
        type: string
      flag_id:
        type: string
      operation_type:
        enum:
        - edit_post
        - close_post
        - delete_post
        - unlist_post
        - ignore_report
        type: string
      tags:
        items:
          $ref: '#/definitions/schema.TagItem'
        type: array
      title:
        maxLength: 150
        minLength: 6
        type: string
    required:
    - flag_id
    - operation_type
    type: object
...
  schema.TagItem:
    properties:
      display_name:
        description: display_name
        maxLength: 35
        type: string
      original_text:
        description: original text
        type: string
      slug_name:
        description: slug_name
        maxLength: 35
        type: string
    type: object
  schema.TagResp:
    properties:
      display_name:
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
    type: object
...
  schema.UpdateCommentReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        description: whether user can delete it
        type: string
      comment_id:
        description: comment id
        type: string
      original_text:
        description: original comment content
        maxLength: 600
        minLength: 2
        type: string
    required:
    - comment_id
    - original_text
    type: object
...
  schema.UserBasicInfo:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      id:
        type: string
      language:
        type: string
      location:
        type: string
      rank:
        type: integer
      status:
        type: string
      suspended_until:
        type: integer
      username:
        type: string
      website:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/comment:
    delete:
      consumes:
      - application/json
      description: remove comment
      parameters:
      - description: comment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RemoveCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: remove comment
      tags:
      - Comment
    get:
      description: get comment by id
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetCommentResp'
                        type: array
                    type: object
              type: object
      summary: get comment by id
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: add comment
      parameters:
      - description: comment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AddCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetCommentResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: add comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: update comment
      parameters:
      - description: comment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update comment
      tags:
      - Comment
  /answer/api/v1/comment/page:
    get:
      description: get comment page
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: object id
        in: query
        name: object_id
        required: true
        type: string
      - description: query condition
        enum:
        - vote
        in: query
        name: query_cond
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetCommentResp'
                        type: array
                    type: object
              type: object
      summary: get comment page
      tags:
      - Comment
...
  /answer/api/v1/personal/comment/page:
    get:
      description: user personal comment list
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: username
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetCommentPersonalWithPageResp'
                        type: array
                    type: object
              type: object
      summary: user personal comment list
      tags:
      - Comment
...
  /answer/api/v1/report:
    post:
      consumes:
      - application/json
      description: add report <br> source (question, answer, comment, user)
      parameters:
      - description: report
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AddReportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: add report
      tags:
      - Report
  /answer/api/v1/report/review:
    put:
      consumes:
      - application/json
      description: review report
      parameters:
      - description: flag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.ReviewReportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: review report
      tags:
      - Report
  /answer/api/v1/report/unreviewed/post:
    get:
      consumes:
      - application/json
      description: get unreviewed report post page
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetReportListPageResp'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: get unreviewed report post page
      tags:
      - Report
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"